[FEATURE] Move the parameters.yml outside of the vendor/folder (#188)
parameters.yml will contain installation-specific data and must not be
overwritten on a composer update. Hence, it needs to reside in the
Configuration/ folder of the application, not in the Configuration/
folder of the core package (will will get overwritten on an update).

Also automatically create the file on a composer install/update
(if the file does not exist yet).

This change is a prerequisite to switching to the Symfony Doctrine
integration.