[FEATURE] Model and repository for subscriber lists
Also rename "list" to "subscriber list" in the domain model. (We cannot use
just "list" as this is a reserved keyword in PHP.)

Also polish the layout of the SubscriberList associations in the graphic
domain model.

This commit does not contain the back association
Administrator -> SubscriberList yet. It will come when it is needed
(either as an association on the model or as a findBy* method in the
repository).