[BUGFIX] Always use UTC as token creation date time zone (#234)
The creation date is saved in the database as a UNIX timestamp
and thus does not contain a time zone. This causes the timestamp
to result in different dates depending on the time zone set in
the PHP configuration, leading to test failures on some systems.

Now the timestamp will always be interpreted to be in the UTC
timezone.